#4201 Add proper qr code icon
